requirements.txt
django
djangorestframework
django-cors-headers
pyyaml
requests
